requirements.txt
pandas
numpy
Werkzeug <= 2.1.2   
flask == 2.1.3   
flask-restx >= 0.5.1 
flask-sqlalchemy==2.5.1
python-dotenv
psycopg2
pytest
mock-alchemy
requests